backend/requirements.txt
fastapi
uvicorn
pydantic
pinecone-client
openai
langchain
langchain-community
sentence-transformers
tiktoken
pygithub
python-dotenv
